Updated photo upload, Updated CBv3 global header
Photo upload has been updated, now the work-flow of photo uploading is
same video uploading and as multiserver. Added new mode in photo upload
called 'process',  after uploading this mode will generate the photo
files.

Javascript for photo upload and the HTML block is now separated. HTML
block is a part of template which should be unique to every template or
if template does not have one, load it from styles/global. All
javascript is in modules/uploader/plupload/plupload.upload_photo.html,
because we need smarty to configure plupload correctly we are using HTML
File. On photo_upload.php, we register an anchor on 'cb_head' to load
javascript code.